\name{AlignedBases-methods}
\docType{methods}
\alias{AlignedBases-methods}
\title{AlignedBases (bamAlignment)}
\description{
Retrieving of aligned bases from alignment.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlignment")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
