\name{bamClose-methods}
\docType{methods}
\alias{bamClose-methods}
\alias{bamClose,bamReader-method}
\alias{bamClose,bamWriter-method}
\title{ bamClose(bamReader),bamClose(bamWriter) }
\description{Closing open file connections for bamReader an bamWriter.}
\section{Methods}{
\describe{
\item{\code{signature(object = "bamReader")}}{Closes connection of bamReader to BAM-file.}
\item{\code{signature(object = "bamWriter")}}{Closes connnection of bamWriter to BAM-file.}
}}

\keyword{methods}
\keyword{bamClose}
