\name{bamOpen-methods}
\docType{methods}
\alias{bamOpen-methods}
\alias{bamOpen,bamReader-method}
\alias{bamOpen,bamWriter-method}
\title{bamOpen (bamReader, bamWriter)}
\description{
Generic function for opening file connection for bamReader and bamWriter objects.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamReader")}}{
Opens connection to BAM-file for bamReader.
}

\item{\code{signature(object = "bamWriter")}}{
Opens connection to BAM-file for bamReader.
}
}}
\keyword{methods}
