\name{bamSave-methods}
\docType{methods}
\alias{bamSave-methods}
\alias{bamSave,bamWriter-method}
\title{bamSave (bamReader)}
\description{
Saves content of bamAlignment object to open BAM file.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamWriter")}}{
%%  ~~describe this method here~~
An open instance of class bamWriter
}
}}
\keyword{methods}
\keyword{bamSave}
