\name{cigarSize-methods}
\docType{methods}
\alias{cigarSize-methods}
\title{Returns cigar data from bamAlign}
\description{Returns cigar data from bamAlign.}
\section{Methods}{
\describe{
\item{\code{signature(object = "bamAlign")}}{An instance of class bamAlign.}
}}
\keyword{methods}
\keyword{bamAlign}
\keyword{cigarSize}
