\name{createIndex-methods}
\docType{methods}
\alias{createIndex-methods}
\title{Creating Index file for sorted BAM file}
\description{ \code{createIndex} Creating Index file for sorted BAM file}
\section{Methods}{
\describe{\item{\code{signature(object = "bamReader")}}{A bamReader object.}}
}
\keyword{methods}
\keyword{createIndex}
