\name{Duplicate<--methods}
\docType{methods}
\alias{Duplicate<--methods}
\title{ Duplicate<-: StandardGeneric}
\description{
Writing accessor for Duplicate attribute.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlignment")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{Duplicate<-}
