\name{failedQC-methods}
\docType{methods}
\alias{failedQC-methods}
\alias{failedQC<--methods}
\title{Reading and writing \code{failedQC} value in bamAlign}
\description{Reading and writing \code{failedQC} value in bamAlign.}
\section{Methods}{
\describe{
\item{\code{signature(object = "bamAlign")}}{
An instance of class bamAlign.
}
}}
\keyword{methods}
\keyword{failedQC}
\keyword{bamAlign}
