\name{FailedQC<--methods}
\docType{methods}
\alias{FailedQC<--methods}
\title{FailedQC (bamAlignment)}
\description{
Writing accessor for FailedQC attribute of bamAlignment object
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlignment")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{FailedQC}
