\name{filename-methods}
\docType{methods}
\alias{filename}
\alias{filename-methods}
\title{Reading \code{filename} from bamReader or bamWriter}
% - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - %
% Descriptions
% - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - %
\description{
    Reading \code{filename} from bamReader or bamWriter
}
% - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - %
% Methods
% - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - %
\section{Methods}{
\describe{
    \item{\code{signature(object = "bamReader")}}{An instance of class bamReader.}
    \item{\code{signature(object = "bamWriter")}}{An instance of bamWriter.}
}}
\keyword{methods}
\keyword{filename}
\keyword{bamReader}
\keyword{bamWriter}
