\name{filename<--methods}
\docType{methods}
\alias{filename<--methods}
\title{filename writing accessor (bamReader, bamWriter)}
\description{
 \code{filename(bamReader)<-"new_filename"}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamReader")}}{
%%  ~~describe this method here~~
Reads filename from bamReader object.
}

\item{\code{signature(object = "bamWriter")}}{
%%  ~~describe this method here~~
Reads filename from bamWriter object.
}
}}
\keyword{methods}
\keyword{filename}
