\name{firstInPair-methods}
\docType{methods}
\alias{firstInPair-methods}
\alias{firstInPair<--methods}
\alias{secondInPair-methods}
\alias{secondInPair<--methods}
\title{Reads or writes information about firstInPair- (secondInPair-) status of bamAlign.}
\description{
Reads or writes information about firstInPair- (secondInPair-) status of bamAlign. FirstInPair and secondInPair status refers to paired end reads.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlign")}}{
An instance of class bamAlign.
}
}}
\keyword{methods}
\keyword{fistInPair}
\keyword{secondInPair}
\keyword{bamAlign}
