\name{flag-methods}
\docType{methods}
\alias{flag-methods}
\alias{flag<--methods}
\title{Reading and writing \code{flag} in bamAlign}
\description{
Reading and writing \code{flag} in bamAlign.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlign")}}{
An instance of class bamAlign.
}
}}
\keyword{methods}
\keyword{flag}
