\name{getHeaderText-methods}
\docType{methods}
\alias{getHeaderText-methods}
\title{ \code{getHeaderText} Assembling member data into Header-Text}
\description{
\code{getHeaderText} Assembling member data into Header-Text. 
}
\section{Methods}{
\describe{
\item{\code{signature(object = "bamHeader")}}{
 An object of class bamHeader
}
}}
\keyword{methods}
\keyword{getHeaderText}
