\name{getNext-methods}
\docType{methods}
\alias{getNext-methods}
\title{getNext (bamAlignment)}
\description{
Retrieving next alignment from BAM-file without creating a new instance of bamAlignment.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlignment")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{getNext}
