\name{GetNextAlignment-methods}
\docType{methods}
\alias{GetNextAlignment-methods}
\title{GetNextAlignmenet (bamReader)}
\description{
Reads next alignment from bamReader.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamReader")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{GetNextAlignment}
