\name{getPrevAlign-methods}
\docType{methods}
\alias{getPrevAlign-methods}
\title{Retrieving previous align from bamRange}
\description{
 Returns previous bamAlign object from bamRange.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamRange")}}{
An instance of class bamRange.
}
}}
\keyword{methods}
\keyword{getPrevAlign}
\keyword{bamRange}
