\name{getRefCount-methods}
\docType{methods}
\alias{getRefCount-methods}
\title{Returns number of reference sequences in BAM file}
\description{
Returns number of reference sequences in BAM file.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamReader")}}{
An instance of clas bamReader
}
}}
\keyword{methods}
\keyword{getRefCount}
\keyword{bamReader}
