\name{getRefData-methods}
\docType{methods}
\alias{getRefData-methods}
\title{Retrieving a table with all reference sequences from a bamReader object.}
\description{Returns a data.frame with all references and an additional ID column.}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamReader")}}{
An object of class bamReader.
}
}}
\keyword{methods}
\keyword{getRefData}
