\name{index.initialized-methods}
\docType{methods}
\alias{index.initialized-methods}
\title{index.initialized(bamReader)}
\description{
The method checks, if the bamReader object contains a valid BAM-index.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamReader")}}{
	An instance of class bamReader
}

}}
\keyword{methods}
\keyword{indes.initialized}
