\name{insertPastCurrent-methods}
\docType{methods}
\alias{insertPastCurrent-methods}
\alias{insertPreCurrent-methods}
\title{Inserting bamAlign before or past the current Align into bamRange}
\description{
Inserting bamAlign before or past the current Align into bamRange.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamRange")}}{
An instance of class bamRange.
}
}}
\keyword{methods}
\keyword{insertPastCurent}
\keyword{insertPreCurrent}
