\name{InsertSize-methods}
\docType{methods}
\alias{InsertSize-methods}
\title{InsertSize (bamAlignment)}
\description{
Returns mate-pair insert size.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlignment")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{InsertSize}
