\name{isValid-methods}
\docType{methods}
\alias{isValid-methods}
\title{isValid (bamAlignment)}
\description{
Returns TRUE when bamAlignment is 'valid'.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlignment")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{isValid}
