\name{Mapped-methods}
\docType{methods}
\alias{Mapped-methods}
\title{Mapped (bamAlignment)}
\description{
Reading accessor for 'Mapped' attribute of bamAlignment objects.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlignment")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{Mapped}
