\name{Mapped<--methods}
\docType{methods}
\alias{Mapped<--methods}
\title{Mapped (bamAlignment)}
\description{
Writing accessor for 'Mapped' attribute of bamAlignment objects.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlignment")}}{
%%  ~~describe this method here~~
An instance of class bamAlignment
}
}}
\keyword{methods}
\keyword{Mapped}
