\name{mapQuality-methods}
\docType{methods}
\alias{mapQuality-methods}
\title{Returns quality value for BAM alignment.}
\description{
Returns quality value for BAM alignment. According to SAM file format specification, this value represents a measure for the probability that the given mapping position is wrong.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlign")}}{
An instance of class bamAlign.
}
}}
\keyword{methods}
\keyword{mapQuality}
\keyword{bamAlign}
