\name{MatePosition-methods}
\docType{methods}
\alias{MatePosition-methods}
\title{MatePosition (bamAlignment}
\description{
Reading accessor for MatePosition attribute of bamAlignment objects.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlignment")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{MatePosition}
