\name{MateRefID-methods}
\docType{methods}
\alias{MateRefID-methods}
\title{MateRefID (bamAlignment)}
\description{
RefID for mate pair alignment.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlignment")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{MateRefID}
