\name{mateReverseStrand-methods}
\docType{methods}
\alias{mateReverseStrand-methods}
\alias{mateReverseStrand<--methods}
\title{Retrieves or writes value for Strand information of mate pair align}
\description{
Retrieves or writes value for Strand information of mate pair align.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlign")}}{
An instance of class bamAlign.
}
}}
\keyword{methods}
\keyword{mateReverseStrand}
\keyword{bamAlign}
