\name{MateReverseStrand<--methods}
\docType{methods}
\alias{MateReverseStrand<--methods}
\title{MateReverseStrand (bamAlignment)}
\description{
Writing accessor for MateReverseStrand attribute.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlignment")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{MateReverseStrand<-}
