\name{mateUnmapped-methods}
\docType{methods}
\alias{mateUnmapped-methods}
\alias{mateUnmapped<--methods}
\title{Retrieves or writes information about mapping status of mate pair align}
\description{
Retrieves or writes information about mapping status of mate pair align.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlign")}}{
An instance of class bamAlign
}
}}
\keyword{methods}
\keyword{mateUnmapped}
\keyword{bamAlign}
