\name{name-methods}
\docType{methods}
\alias{name-methods}
\title{Retrieving read identifier from bamAlign}
\description{
Retrieving read identifier from bamAlign.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlign")}}{
An instance of class bamAlign.
}
}}
\keyword{methods}
\keyword{name}
\keyword{bamAlign}
