\name{name<--methods}
\docType{methods}
\alias{name<--methods}
\title{name (bamAlignment)}
\description{
Writing accessor for name property of bamAlignment objects.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlignment")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{name<-}
