\name{nGapAligns}
\Rdversion{1.1}
\docType{class}
\alias{nGapAligns}
\alias{nGapAligns-methods}
\alias{nGapAligns,gapList-method}
\alias{nGapAligns,gapSiteList-method}
\alias{nGapAligns,bamGapList-method}
\title{nGapAligns}
\description{
The function provides a deprecated accessor to the total number of align-gaps counted in various objects. The new accessor for this slot is named \code{nAlignGaps}. 
}
\author{Wolfgang Kaisers}
\examples{
# No examples
}
\keyword{gapList}
