\name{Paired-methods}
\docType{methods}
\alias{Paired-methods}
\title{Paired (bamAlignment)}
\description{
Reading accessor for Paired attribute of bamAlignment.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlignment")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{Paired}
