\name{pcrORopt_duplicate-methods}
\docType{methods}
\alias{pcrORopt_duplicate-methods}
\alias{pcrORopt_duplicate<--methods}
\title{Retrieves or writes if align is PCR or optical duplicate.}
\description{
Retrieves or writes if align is PCR or optical duplicate.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlign")}}{
An instance of class bamAlign.
}
}}
\keyword{methods}
\keyword{pcrORopt_duplicate}
\keyword{bamAlign}
