\name{pop_front-methods}
\docType{methods}
\alias{pop_front-methods}
\alias{pop_back-methods}
\title{Removes first (last) bamAlign from bamRange}
\description{
Removes first (last) bamAlign from bamRange.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamRange")}}{
An instance of class bamAlign.
}
}}
\keyword{methods}
\keyword{pop_front}
\keyword{pop_back}
\keyword{bamRange}
