\name{position-methods}
\docType{methods}
\alias{position-methods}
\title{Reads position data member from bamAlign}
\description{
Reads position data member from bamAlign.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlign")}}{
An instance of class bamAlign.
}
}}
\keyword{methods}
\keyword{position}
\keyword{bamAlign}
