\name{PrimaryAlignment-methods}
\docType{methods}
\alias{PrimaryAlignment-methods}
\title{PrimaryAlignment (bamAlignment)}
\description{
Reading accessor for PrimaryAlignment attribute of bamAlignment.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlignment")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{PrimaryAlignment}
