\name{properPair-methods}
\docType{methods}
\alias{properPair-methods}
\alias{properPair<--methods}
\title{Reads or writes the properPair status of bamAlign}
\description{
Reads or writes the properPair status of bamAlign. The properPair status describes whether a paired sequenced read is also mapped in a proper pair.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlign")}}{
An instance of class bamAlign.
}
}}
\keyword{methods}
\keyword{properPair}
\keyword{paired}
