\name{ProperPair<--methods}
\docType{methods}
\alias{ProperPair<--methods}
\title{ProperPair (bamAlignemnt)}
\description{
Writing accessor for ProperPair attribute of bamAlignment.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlignment")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{ProperPair<-}
