\name{push_back-methods}
\docType{methods}
\alias{push_back-methods}
\alias{push_front-methods}
\title{Appends bamAlign before the first or behind the last element of bamRange}
\description{
 Appends bamAlign before the first or behind the last element of bamRange.
}
\section{Methods}{
\describe{
\item{\code{signature(object = "bamRange")}}{
An instance of class bamRange.
}
\item{\code{signature(object = "bamRange")}}{
An instance of class bamRange.
}
}}
\keyword{methods}
\keyword{push_back}
\keyword{push_front}
