\name{Qualities-methods}
\docType{methods}
\alias{Qualities-methods}
\title{Qualities (bamAlignment)}
\description{
Reading accessor for Qualities attribute of bamAlignment.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlignment")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{Qualities}
