\name{readBases-methods}
\docType{methods}
\alias{readBases-methods}
\title{Returns the sequence of nucleobases of the aligned read from a bamAlign}
\description{
Returns the sequence of nucleobases of the aligned read from a bamAlign.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlign")}}{
An instance of class bamAlign.
}
}}
\keyword{methods}
\keyword{readBases}
\keyword{bamAlign}
