\name{refID-methods}
\docType{methods}
\alias{refID-methods}
\title{refID (bamAlignment)}
\description{
Reading accessor for refID attribute of bamAlingment
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlignment")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{refID}
