\name{reverseStrand-methods}
\docType{methods}
\alias{reverseStrand-methods}
\alias{reverseStrand<--methods}
\title{Retrieves strand information from bamAlign}
\description{
Retrieves strand information from bamAlign.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlign")}}{
An instance of class bamAlign
}
}}
\keyword{methods}
\keyword{reverseStrand}
\keyword{bamAlign}
