\name{Rewind-methods}
\docType{methods}
\alias{Rewind-methods}
\title{Rewind (bamReader)}
\description{
Resets the reading position of bamReader to beginning of file. The next call to GetNextAlignment then retrieves the first alignment in the opened BAM-file.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamReader")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
