\name{samHeader-methods}
\docType{methods}
\alias{samHeader-methods}
\title{samHeader (bamReader)}
\description{
Reading accessor for SAM-header data in bamWriter and bamReader objects.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamWriter")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{samHeader}
