\name{samHeader<--methods}
\docType{methods}
\alias{samHeader<--methods}
\title{samHeader<-}
\description{
Writing accessor for SAM-header data in bamWriter.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamWriter")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{samHeader<-}
