\name{secondaryAlign-methods}
\docType{methods}
\alias{secondaryAlign-methods}
\alias{secondaryAlign<--methods}
\title{Reads or writes information about the secondaryAlign status of bamAlign}
\description{
Reads or writes information about the secondaryAlign status of bamAlign. The SAM file format specification says that this flag marks the alignment to be excluded from analysis.

}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlign")}}{
An instance of class bamAlign.
}
}}
\keyword{methods}
\keyword{secondaryAlign}
\keyword{bamAlign}
