\name{SecondMate<--methods}
\docType{methods}
\alias{SecondMate<--methods}
\title{SecondMate (bamAlignment)}
\description{
Writing accessor for SecondMate attribute of bamAlignment.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlignment")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{SecondMate<-}
