\name{unmapped-methods}
\docType{methods}
\alias{unmapped-methods}
\alias{unmapped<--methods}
\title{Reads or writes mapping status of bamAlign}
\description{Reads or writes mapping status of bamAlign. According to the SAM file format specification, this is the only reliable flag which determines whether a read is unaligned.}
\section{Methods}{
\describe{
\item{\code{signature(object = "bamAlign")}}{An instance of class bamAlign.}
}}
\keyword{methods}
\keyword{unmapped}
