\name{windBack-methods}
\docType{methods}
\alias{windBack-methods}
\title{Resets the current align pointer of a bamRange to the position before the first align}
\description{
Resets the current align pointer of a bamRange to the position before the first align. The next call of getNextAlign then returns the first align of the bamRange container.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamRange")}}{
An instance of class bamRange.
}
}}
\keyword{methods}
\keyword{windBack}
\keyword{bamRange}
