\name{writeCurrentAlign-methods}
\docType{methods}
\alias{writeCurrentAlign-methods}
\title{Replaces current align of bamRange with the given bamAlign}
\description{
Replaces current align of bamRange with the given bamAlign.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamRange")}}{
An object of class bamRange.
}
}}
\keyword{methods}
\keyword{bamRange}
