\name{bamClose-methods}
\docType{methods}
\alias{bamClose-methods}
\alias{bamClose,bamReader-method}
\alias{bamClose,bamWriter-method}
\title{bamClose (bamReader, bamWriter)}
\description{
bamClose is a generic function for closing file connections for bamReader and bamWriter objects.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamReader")}}{
%%  ~~describe this method here~~
Closes connection to BAM-file for bamReader object.
}

\item{\code{signature(object = "bamWriter")}}{
%%  ~~describe this method here~~
Closes connection to BAM-file for bamWriter object.
}
}}
\keyword{methods}
