\name{bamClose-methods}
\docType{methods}
\alias{bamClose}
\alias{bamClose-methods}
\title{bamClose(bamReader), bamClose(bamWriter)}
\description{Closing open file connections for bamReader an bamWriter.}
\section{Methods}{
\describe{
\item{\code{signature(object = "bamReader")}}{Closes connection of bamReader to BAM-file.}
\item{\code{signature(object = "bamWriter")}}{Closes connnection of bamWriter to BAM-file.}
}}
\keyword{methods}
\keyword{bamClose}
