\name{bamClose-methods}
\docType{methods}
\alias{bamClose}
\alias{bamClose-methods}
\title{bamClose(bamReader), bamClose(bamWriter): Closing of file connections.}
\description{Closing open file connections for bamReader an bamWriter.}
% - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - %
% Methods
% - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - %
\section{Methods}{
\describe{
    \item{\code{signature(object = "bamReader")}}{
        Closes connection of bamReader to BAM-file.}
    %
    \item{\code{signature(object = "bamWriter")}}{
        Closes connnection of bamWriter to BAM-file.}
}}
\keyword{methods}
\keyword{bamClose}
