\name{bamSave-methods}
\docType{methods}
\alias{bamSave-methods}
\alias{bamSave,bamRange-method}
\alias{bamSave,bamWriter-method}
\title{Writes bamAligns to an open bamWriter instance}
\description{
Write a single bamAlign or aligns contained in a bamRange to an open BAM-file (represented by a bamWriter object)
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamRange")}}{
An instance of class bamRange
}
}
}
\keyword{methods}
\keyword{bamAlign}
\keyword{bamRange}
\keyword{bamWriter}
