\name{CigarData-methods}
\docType{methods}
\alias{CigarData-methods}
\title{CigarData (bamAlignment)}
\description{
Returns cigar-data from bamAlignment object as data.frame.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlignment")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{CigarData}
