\name{CreateIndex-methods}
\docType{methods}
\alias{CreateIndex-methods}
\title{CreateIndex (bamReader)}
\description{
Creates index file (filename.bam.bai) for open bamReader object.
}
\section{Methods}{
\describe{
\item{\code{signature(object = "bamReader")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{CreateIndex}
