\name{filename-methods}
\docType{methods}
\alias{filename-methods}
\alias{filename,bamReader-method}
\alias{filename,bamWriter-method}
\title{Reading \code{filename} from bamReader or bamWriter}
\description{
 Reading \code{filename} from bamReader or bamWriter
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamReader")}}{
An instance of class bamReader
}

\item{\code{signature(object = "bamWriter")}}{
An instance of bamWriter
}
}}
\keyword{methods}
\keyword{filename}
\keyword{bamReader}
\keyword{bamWriter}
