\name{getNextAlign-methods}
\docType{methods}
\alias{getNextAlign-methods}
\title{\code{getNextAlign}: Retrieving next align from bamReader or bamRange}
\description{
Returns an object of class bamAlign from bamReader or bamRange.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamRange")}}{
An instance of class bamRange.
}

\item{\code{signature(object = "bamReader")}}{
An instance of class bamReader
}
}}
\keyword{methods}
\keyword{getNextAlign}
\keyword{bamReader}
\keyword{bamRange}
