\name{insertSize-methods}
\docType{methods}
\alias{insertSize-methods}
\title{Returns value for insertSize from bamAlign}
\description{
Returns value for insertSize from bamAlign
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlign")}}{
An instance of class bamAlign
}
}}
\keyword{methods}
\keyword{insertSize}
\keyword{bamAlign}
