\name{isOpen-methods}
\docType{methods}
\alias{isOpen-methods}
\alias{isOpen,bamReader-method}
\alias{isOpen,bamWriter-method}
\title{ isOpen(bamReader),isOpen(bamWriter) }
\description{
Checks whether the given bamReader or bamWriter represents an opened connection to a BAM-file.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamReader")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "bamWriter")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{isOpen}
