\name{matePosition-methods}
\docType{methods}
\alias{matePosition-methods}
\title{Returns position of mate pair align}
\description{
Returns position of mate pair align.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlign")}}{
An instance of class bamAlign
}
}}
\keyword{methods}
\keyword{matePosition}
