\name{MateMapped<--methods}
\docType{methods}
\alias{MateMapped<--methods}
\title{MateMapped (bamAlignment}
\description{
Writing accessor for MateMapped attribute of bamAlignment objects.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlignment")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{MateMapped<-}
