\name{mateRefID-methods}
\docType{methods}
\alias{mateRefID-methods}
\title{Returns refID of mate pair align}
\description{
 Returns refID of mate pair align.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlign")}}{
An instance of class bamAlign
}
}}
\keyword{methods}
\keyword{mateRefID}
