\name{name-methods}
\docType{methods}
\alias{name-methods}
\title{name (bamAlignment)}
\description{
Reading accessor for name attribute of bamAlignment.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlignment")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{name}
