\name{paired-methods}
\docType{methods}
\alias{paired-methods}
\alias{paired<--methods}
\title{Retrieves or writes if align belongs to a paired-end align}
\description{
Retrieves or writes if align belongs to a paired-end align.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlign")}}{
An instance of class bamAlign.
}
}}
\keyword{methods}
\keyword{paired}
\keyword{bamAlign}
