\name{position-methods}
\docType{methods}
\alias{position-methods}
\title{position (bamAlignment)}
\description{
Reading accessor for position attribute of bamAlignment.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlignment")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{position}
