\name{ProperPair-methods}
\docType{methods}
\alias{ProperPair-methods}
\title{ProperPair (bamAlignment)}
\description{
Reading Accessor for ProperPair attribute of bamAlignment.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlignment")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{ProperPair}
