\name{refID-methods}
\docType{methods}
\alias{refID-methods}
\title{Returns refID from bamAlign}
\description{
Returns refID from bamAlign.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "bamAlign")}}{
An instance of class bamAlign.
}
}}
\keyword{methods}
\keyword{refID}
\keyword{bamAlign}
